app/requirements.txt
asgiref==3.5.2
Django==4.1.3
pip==21.1.2
setuptools==57.0.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.36.2
pillow
djangorestframework==3.14.0
django-environ
drf-spectacular==0.24.2
uwsgi


